refactor: import interfaces from dataTypes.ts
