+++ src/hooks/useColumnVisibility.tsx
@@ -1,24 +1,7 @@
 // import { useState, useMemo } from "react";
 import { useState } from "react";
 
-// --- Interfaces
-export interface ColumnVisibility {
-  myID: boolean;
-  Name: boolean;
-  CreatedTime: boolean;
-  EditedTime: boolean;
-  NotedTime: boolean;
-  CreatedEnd: boolean;
-  PublishedStart: boolean;
-  PublishedEnd: boolean;
-  Area: boolean;
-  Source: boolean;
-  Link: boolean;
-  Type: boolean;
-  Tags: boolean;
-  PageURL: boolean;
-  // pageContent: boolean;
-}
+import { ColumnVisibility } from "../utils/dataTypes";
 // --- Default and Preset Column Visibility Settings ---
 export const defaultColumnVisibility: ColumnVisibility = {
   myID: true,
